Fuzz: use spdmlib-test testing utils
