+++ fuzz-target/fuzzlib/Cargo.toml
@@ -8,16 +8,12 @@ edition = "2018"
 
 [dependencies]
 afl = { version = "=0.12.12", optional = true }
-spdm-emu = { path = "../../test/spdm-emu", default-features = false }
 spdmlib = { path = "../../spdmlib", default-features = false, features=["spdm-ring"] }
-codec = { path = "../../codec" }
-mctp_transport = { path = "../../mctp_transport" }
-pcidoe_transport = { path = "../../pcidoe_transport" }
 simple_logger = "1.11.0"
 log = "0.4.13"
-bytes = { version="1", default-features=false }
 ring = { version = "0.16.20" }
 flexi_logger = "0.18.0"
+spdmlib-test = { path = "../../test/spdmlib-test" }
 
 [features]
 default = ["hashed-transcript-data", "afl"]
