--- fuzz-target/fuzzlib/src/requesterlib.rs
@@ -1,255 +1,0 @@
-// Copyright (c) 2022 Intel Corporation
-//
-// SPDX-License-Identifier: BSD-2-Clause-Patent
-
-use super::*;
-use spdmlib::protocol::*;
-
-pub fn req_create_info() -> (common::SpdmConfigInfo, common::SpdmProvisionInfo) {
-    let config_info = common::SpdmConfigInfo {
-        spdm_version: [
-            SpdmVersion::SpdmVersion10,
-            SpdmVersion::SpdmVersion11,
-            SpdmVersion::SpdmVersion12,
-        ],
-        req_capabilities: SpdmRequestCapabilityFlags::CERT_CAP
-        | SpdmRequestCapabilityFlags::CHAL_CAP
-        | SpdmRequestCapabilityFlags::ENCRYPT_CAP
-        | SpdmRequestCapabilityFlags::MAC_CAP
-        //| SpdmRequestCapabilityFlags::MUT_AUTH_CAP
-        | SpdmRequestCapabilityFlags::KEY_EX_CAP
-        | SpdmRequestCapabilityFlags::PSK_CAP
-        | SpdmRequestCapabilityFlags::ENCAP_CAP
-        | SpdmRequestCapabilityFlags::HBEAT_CAP
-        | SpdmRequestCapabilityFlags::KEY_UPD_CAP, // | SpdmRequestCapabilityFlags::HANDSHAKE_IN_THE_CLEAR_CAP
-        // | SpdmRequestCapabilityFlags::PUB_KEY_ID_CAP
-        req_ct_exponent: 0,
-        measurement_specification: SpdmMeasurementSpecification::DMTF,
-        base_asym_algo: if USE_ECDSA {
-            SpdmBaseAsymAlgo::TPM_ALG_ECDSA_ECC_NIST_P384
-        } else {
-            SpdmBaseAsymAlgo::TPM_ALG_RSASSA_3072
-        },
-        base_hash_algo: SpdmBaseHashAlgo::TPM_ALG_SHA_384,
-        dhe_algo: SpdmDheAlgo::SECP_384_R1,
-        aead_algo: SpdmAeadAlgo::AES_256_GCM,
-        req_asym_algo: SpdmReqAsymAlgo::TPM_ALG_RSAPSS_2048,
-        key_schedule_algo: SpdmKeyScheduleAlgo::SPDM_KEY_SCHEDULE,
-        ..Default::default()
-    };
-
-    let mut peer_root_cert_data = SpdmCertChainData {
-        ..Default::default()
-    };
-    let crate_dir = get_test_key_directory();
-    let ca_file_path = if USE_ECDSA {
-        "test_key/ecp384/ca.cert.der"
-    } else {
-        "test_key/rsa3072/ca.cert.der"
-    };
-    let ca_cert = std::fs::read(crate_dir.join(ca_file_path)).expect("unable to read ca cert!");
-    let inter_file_path = if USE_ECDSA {
-        "test_key/ecp384/inter.cert.der"
-    } else {
-        "test_key/rsa3072/inter.cert.der"
-    };
-    let inter_cert =
-        std::fs::read(crate_dir.join(inter_file_path)).expect("unable to read inter cert!");
-    let leaf_file_path = if USE_ECDSA {
-        "test_key/ecp384/end_responder.cert.der"
-    } else {
-        "test_key/rsa3072/end_responder.cert.der"
-    };
-    let leaf_cert =
-        std::fs::read(crate_dir.join(leaf_file_path)).expect("unable to read leaf cert!");
-
-    let ca_len = ca_cert.len();
-    let inter_len = inter_cert.len();
-    let leaf_len = leaf_cert.len();
-    log::info!(
-        "total cert size - {:?} = {:?} + {:?} + {:?}",
-        ca_len + inter_len + leaf_len,
-        ca_len,
-        inter_len,
-        leaf_len
-    );
-    peer_root_cert_data.data_size = (ca_len) as u16;
-    peer_root_cert_data.data[0..ca_len].copy_from_slice(ca_cert.as_ref());
-
-    let provision_info = common::SpdmProvisionInfo {
-        my_cert_chain_data: [None, None, None, None, None, None, None, None],
-        my_cert_chain: [None, None, None, None, None, None, None, None],
-        peer_root_cert_data: Some(peer_root_cert_data),
-    };
-
-    (config_info, provision_info)
-}
-
-pub fn certificata_data() -> [[u8; 528]; 4] {
-    [
-        [
-            0x1, 0x0, 0x1, 0x0, 0x84, 0x0, 0x0, 0x0, 0x11, 0x2, 0x0, 0x0, 0x0, 0x2, 0x8, 0x4, 0x8,
-            0x6, 0x0, 0x0, 0x5a, 0x64, 0xb3, 0x8b, 0x5d, 0x5f, 0x4d, 0xb3, 0x5f, 0xb2, 0xaa, 0x1d,
-            0x46, 0x9f, 0x6a, 0xdc, 0xca, 0x7f, 0xac, 0x85, 0xbe, 0xf0, 0x84, 0x10, 0x9c, 0xcd,
-            0x54, 0x9, 0xf0, 0xab, 0x38, 0x3a, 0xaa, 0xf7, 0xa6, 0x2e, 0x3b, 0xd7, 0x81, 0x2c,
-            0xea, 0x24, 0x7e, 0x14, 0xa9, 0x56, 0x9d, 0x28, 0x30, 0x82, 0x1, 0xcf, 0x30, 0x82, 0x1,
-            0x56, 0xa0, 0x3, 0x2, 0x1, 0x2, 0x2, 0x14, 0x20, 0x3a, 0xc2, 0x59, 0xcc, 0xda, 0xcb,
-            0xf6, 0x72, 0xf1, 0xc0, 0x1a, 0x62, 0x1a, 0x45, 0x82, 0x90, 0x24, 0xb8, 0xaf, 0x30,
-            0xa, 0x6, 0x8, 0x2a, 0x86, 0x48, 0xce, 0x3d, 0x4, 0x3, 0x3, 0x30, 0x1f, 0x31, 0x1d,
-            0x30, 0x1b, 0x6, 0x3, 0x55, 0x4, 0x3, 0xc, 0x14, 0x69, 0x6e, 0x74, 0x65, 0x6c, 0x20,
-            0x74, 0x65, 0x73, 0x74, 0x20, 0x45, 0x43, 0x50, 0x32, 0x35, 0x36, 0x20, 0x43, 0x41,
-            0x30, 0x1e, 0x17, 0xd, 0x32, 0x31, 0x30, 0x32, 0x30, 0x39, 0x30, 0x30, 0x35, 0x30,
-            0x35, 0x38, 0x5a, 0x17, 0xd, 0x33, 0x31, 0x30, 0x32, 0x30, 0x37, 0x30, 0x30, 0x35,
-            0x30, 0x35, 0x38, 0x5a, 0x30, 0x1f, 0x31, 0x1d, 0x30, 0x1b, 0x6, 0x3, 0x55, 0x4, 0x3,
-            0xc, 0x14, 0x69, 0x6e, 0x74, 0x65, 0x6c, 0x20, 0x74, 0x65, 0x73, 0x74, 0x20, 0x45,
-            0x43, 0x50, 0x32, 0x35, 0x36, 0x20, 0x43, 0x41, 0x30, 0x76, 0x30, 0x10, 0x6, 0x7, 0x2a,
-            0x86, 0x48, 0xce, 0x3d, 0x2, 0x1, 0x6, 0x5, 0x2b, 0x81, 0x4, 0x0, 0x22, 0x3, 0x62, 0x0,
-            0x4, 0x99, 0x8f, 0x81, 0x68, 0x9a, 0x83, 0x9b, 0x83, 0x39, 0xad, 0xe, 0x32, 0x8d, 0xb9,
-            0x42, 0xd, 0xae, 0xcc, 0x91, 0xa9, 0xbc, 0x4a, 0xe1, 0xbb, 0x79, 0x4c, 0x22, 0xfa,
-            0x3f, 0xc, 0x9d, 0x93, 0x3c, 0x1a, 0x2, 0x5c, 0xc2, 0x73, 0x5, 0xec, 0x43, 0x5d, 0x4,
-            0x2, 0xb1, 0x68, 0xb3, 0xf4, 0xd8, 0xde, 0xc, 0x8d, 0x53, 0xb7, 0x4, 0x8e, 0xa1, 0x43,
-            0x9a, 0xeb, 0x31, 0xd, 0xaa, 0xce, 0x89, 0x2d, 0xba, 0x73, 0xda, 0x4f, 0x1e, 0x39,
-            0x5d, 0x92, 0x11, 0x21, 0x38, 0xb4, 0x0, 0xd4, 0xf5, 0x55, 0x8c, 0xe8, 0x71, 0x30,
-            0x3d, 0x46, 0x83, 0xf4, 0xc4, 0x52, 0x50, 0xda, 0x12, 0x5b, 0xa3, 0x53, 0x30, 0x51,
-            0x30, 0x1d, 0x6, 0x3, 0x55, 0x1d, 0xe, 0x4, 0x16, 0x4, 0x14, 0xcf, 0x9, 0xd4, 0x7a,
-            0xee, 0x8, 0x90, 0x62, 0xbf, 0xe6, 0x9c, 0xb4, 0xb9, 0xdf, 0xe1, 0x41, 0x33, 0x1c, 0x3,
-            0xa5, 0x30, 0x1f, 0x6, 0x3, 0x55, 0x1d, 0x23, 0x4, 0x18, 0x30, 0x16, 0x80, 0x14, 0xcf,
-            0x9, 0xd4, 0x7a, 0xee, 0x8, 0x90, 0x62, 0xbf, 0xe6, 0x9c, 0xb4, 0xb9, 0xdf, 0xe1, 0x41,
-            0x33, 0x1c, 0x3, 0xa5, 0x30, 0xf, 0x6, 0x3, 0x55, 0x1d, 0x13, 0x1, 0x1, 0xff, 0x4, 0x5,
-            0x30, 0x3, 0x1, 0x1, 0xff, 0x30, 0xa, 0x6, 0x8, 0x2a, 0x86, 0x48, 0xce, 0x3d, 0x4, 0x3,
-            0x3, 0x3, 0x67, 0x0, 0x30, 0x64, 0x2, 0x30, 0x5a, 0xb4, 0xf5, 0x95, 0x25, 0x82, 0xf6,
-            0x68, 0x3e, 0x49, 0xc7, 0xb4, 0xbb, 0x42, 0x81, 0x91, 0x7e, 0x38, 0xd0, 0x2d, 0xac,
-            0x53, 0xae, 0x8e, 0xb0, 0x51, 0x50, 0xaa, 0xf8, 0x7e, 0xff, 0xc0, 0x30, 0xab, 0xd5,
-            0x8, 0x5b, 0x6, 0xf7, 0xe1, 0xbf, 0x39, 0xd2, 0x3e, 0xae, 0xbf, 0x8e, 0x48, 0x2, 0x30,
-            0x9, 0x75, 0xa8, 0xc0, 0x6f, 0x4f, 0x3c, 0xad, 0x5d, 0x4e, 0x4f, 0xf8, 0x2c, 0x3b,
-            0x39, 0x46, 0xa0, 0xdf, 0x83, 0x8e, 0xb5, 0xd3, 0x61, 0x61, 0x59, 0xbc, 0x39, 0xd7,
-            0xad, 0x68, 0x5e, 0xd, 0x4f, 0x3f, 0xe2, 0xca, 0xc1, 0x74, 0x8f, 0x47, 0x37,
-        ],
-        [
-            0x1, 0x0, 0x1, 0x0, 0x84, 0x0, 0x0, 0x0, 0x11, 0x2, 0x0, 0x0, 0x0, 0x2, 0x8, 0x2, 0x11,
-            0xc8, 0x22, 0x59, 0x6f, 0x64, 0x52, 0x30, 0x82, 0x1, 0xd7, 0x30, 0x82, 0x1, 0x5d, 0xa0,
-            0x3, 0x2, 0x1, 0x2, 0x2, 0x1, 0x1, 0x30, 0xa, 0x6, 0x8, 0x2a, 0x86, 0x48, 0xce, 0x3d,
-            0x4, 0x3, 0x3, 0x30, 0x1f, 0x31, 0x1d, 0x30, 0x1b, 0x6, 0x3, 0x55, 0x4, 0x3, 0xc, 0x14,
-            0x69, 0x6e, 0x74, 0x65, 0x6c, 0x20, 0x74, 0x65, 0x73, 0x74, 0x20, 0x45, 0x43, 0x50,
-            0x32, 0x35, 0x36, 0x20, 0x43, 0x41, 0x30, 0x1e, 0x17, 0xd, 0x32, 0x31, 0x30, 0x32,
-            0x30, 0x39, 0x30, 0x30, 0x35, 0x30, 0x35, 0x39, 0x5a, 0x17, 0xd, 0x33, 0x31, 0x30,
-            0x32, 0x30, 0x37, 0x30, 0x30, 0x35, 0x30, 0x35, 0x39, 0x5a, 0x30, 0x2e, 0x31, 0x2c,
-            0x30, 0x2a, 0x6, 0x3, 0x55, 0x4, 0x3, 0xc, 0x23, 0x69, 0x6e, 0x74, 0x65, 0x6c, 0x20,
-            0x74, 0x65, 0x73, 0x74, 0x20, 0x45, 0x43, 0x50, 0x32, 0x35, 0x36, 0x20, 0x69, 0x6e,
-            0x74, 0x65, 0x72, 0x6d, 0x65, 0x64, 0x69, 0x61, 0x74, 0x65, 0x20, 0x63, 0x65, 0x72,
-            0x74, 0x30, 0x76, 0x30, 0x10, 0x6, 0x7, 0x2a, 0x86, 0x48, 0xce, 0x3d, 0x2, 0x1, 0x6,
-            0x5, 0x2b, 0x81, 0x4, 0x0, 0x22, 0x3, 0x62, 0x0, 0x4, 0x77, 0x1b, 0x24, 0xf6, 0xc6,
-            0x76, 0x1f, 0xb8, 0x30, 0x7, 0x8b, 0xb8, 0xa3, 0x9e, 0xc0, 0x26, 0xc1, 0xea, 0x7d,
-            0xfc, 0x29, 0x7d, 0xe0, 0x59, 0xb2, 0x64, 0x32, 0x75, 0x4a, 0xe3, 0x2, 0x64, 0x3c,
-            0xbc, 0x85, 0x8e, 0xc6, 0xec, 0xef, 0xb0, 0x79, 0xf4, 0xc1, 0xa4, 0xb9, 0xbb, 0x29,
-            0x6b, 0xae, 0xad, 0xf0, 0x7d, 0x63, 0xc6, 0xaf, 0xb3, 0x73, 0x5e, 0x4f, 0x3f, 0xfe,
-            0x89, 0x8a, 0xbb, 0x7d, 0x2b, 0x60, 0x3e, 0x16, 0xba, 0x82, 0xcf, 0xa4, 0x70, 0x4,
-            0x85, 0xc3, 0xa3, 0x3c, 0x5e, 0x6a, 0xa0, 0xef, 0xda, 0xd5, 0x20, 0x30, 0x19, 0xba,
-            0x79, 0x95, 0xb0, 0xc2, 0x7f, 0x4c, 0xdd, 0xa3, 0x5e, 0x30, 0x5c, 0x30, 0xc, 0x6, 0x3,
-            0x55, 0x1d, 0x13, 0x4, 0x5, 0x30, 0x3, 0x1, 0x1, 0xff, 0x30, 0xb, 0x6, 0x3, 0x55, 0x1d,
-            0xf, 0x4, 0x4, 0x3, 0x2, 0x1, 0xfe, 0x30, 0x1d, 0x6, 0x3, 0x55, 0x1d, 0xe, 0x4, 0x16,
-            0x4, 0x14, 0x12, 0xe0, 0x1a, 0x23, 0xc6, 0x23, 0xe4, 0x2, 0x58, 0xb, 0x6, 0xac, 0x90,
-            0xfa, 0x4b, 0x80, 0x3d, 0xc9, 0xf1, 0x1d, 0x30, 0x20, 0x6, 0x3, 0x55, 0x1d, 0x25, 0x1,
-            0x1, 0xff, 0x4, 0x16, 0x30, 0x14, 0x6, 0x8, 0x2b, 0x6, 0x1, 0x5, 0x5, 0x7, 0x3, 0x1,
-            0x6, 0x8, 0x2b, 0x6, 0x1, 0x5, 0x5, 0x7, 0x3, 0x2, 0x30, 0xa, 0x6, 0x8, 0x2a, 0x86,
-            0x48, 0xce, 0x3d, 0x4, 0x3, 0x3, 0x3, 0x68, 0x0, 0x30, 0x65, 0x2, 0x30, 0x3, 0x32,
-            0xb1, 0x8b, 0x20, 0xf4, 0x76, 0xda, 0x8c, 0x83, 0x96, 0x87, 0x55, 0xd9, 0x12, 0x72,
-            0xbd, 0x58, 0x4d, 0xa, 0x37, 0xaf, 0x29, 0x95, 0x1d, 0x36, 0xc4, 0x9e, 0xa5, 0xcd,
-            0xe2, 0x3b, 0xf5, 0xe0, 0x7a, 0x64, 0x36, 0x1e, 0xd4, 0xf1, 0xe1, 0xbb, 0x14, 0x57,
-            0x9e, 0x86, 0x82, 0x72, 0x2, 0x31, 0x0, 0xc0, 0xd6, 0x2, 0x99, 0x50, 0x76, 0x34, 0x16,
-            0xd6, 0x51, 0x9c, 0xc4, 0x86, 0x8, 0x68, 0x94, 0xbf, 0x3c, 0x9, 0x7e, 0x10, 0xe5, 0x62,
-            0x8a, 0xba, 0x48, 0xa, 0xa5, 0xed, 0x1a, 0x6a, 0xf6, 0x3c, 0x2f, 0x4d, 0x38, 0x5d,
-            0x7d, 0x5c, 0x60, 0x63, 0x88, 0x84, 0x5d, 0x49, 0x33, 0xe2, 0xa7, 0x30, 0x82, 0x2,
-            0x22, 0x30, 0x82, 0x1, 0xa8, 0xa0, 0x3, 0x2, 0x1, 0x2, 0x2, 0x1, 0x3, 0x30, 0xa, 0x6,
-            0x8, 0x2a, 0x86, 0x48, 0xce, 0x3d, 0x4, 0x3, 0x3, 0x30, 0x2e,
-        ],
-        [
-            0x1, 0x0, 0x1, 0x0, 0x84, 0x0, 0x0, 0x0, 0x11, 0x2, 0x0, 0x0, 0x0, 0x2, 0x8, 0x0, 0x31,
-            0x2c, 0x30, 0x2a, 0x6, 0x3, 0x55, 0x4, 0x3, 0xc, 0x23, 0x69, 0x6e, 0x74, 0x65, 0x6c,
-            0x20, 0x74, 0x65, 0x73, 0x74, 0x20, 0x45, 0x43, 0x50, 0x32, 0x35, 0x36, 0x20, 0x69,
-            0x6e, 0x74, 0x65, 0x72, 0x6d, 0x65, 0x64, 0x69, 0x61, 0x74, 0x65, 0x20, 0x63, 0x65,
-            0x72, 0x74, 0x30, 0x1e, 0x17, 0xd, 0x32, 0x31, 0x30, 0x32, 0x30, 0x39, 0x30, 0x30,
-            0x35, 0x30, 0x35, 0x39, 0x5a, 0x17, 0xd, 0x32, 0x32, 0x30, 0x32, 0x30, 0x39, 0x30,
-            0x30, 0x35, 0x30, 0x35, 0x39, 0x5a, 0x30, 0x2b, 0x31, 0x29, 0x30, 0x27, 0x6, 0x3, 0x55,
-            0x4, 0x3, 0xc, 0x20, 0x69, 0x6e, 0x74, 0x65, 0x6c, 0x20, 0x74, 0x65, 0x73, 0x74, 0x20,
-            0x45, 0x43, 0x50, 0x32, 0x35, 0x36, 0x20, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x64,
-            0x65, 0x72, 0x20, 0x63, 0x65, 0x72, 0x74, 0x30, 0x76, 0x30, 0x10, 0x6, 0x7, 0x2a, 0x86,
-            0x48, 0xce, 0x3d, 0x2, 0x1, 0x6, 0x5, 0x2b, 0x81, 0x4, 0x0, 0x22, 0x3, 0x62, 0x0, 0x4,
-            0x6c, 0x22, 0x41, 0xdf, 0xb7, 0xe4, 0xd6, 0x8d, 0x53, 0x72, 0x4e, 0x4a, 0x1b, 0x99,
-            0x82, 0xe6, 0x56, 0xd2, 0x2d, 0x97, 0x4b, 0x98, 0x40, 0xa9, 0x99, 0xd6, 0xd, 0xd8,
-            0xe9, 0xa6, 0xfc, 0x74, 0xb9, 0xce, 0x89, 0x48, 0xa7, 0xb5, 0x9, 0xb6, 0x24, 0x49,
-            0xd6, 0x23, 0xb3, 0x5f, 0x3a, 0xf0, 0x99, 0xb0, 0xca, 0x63, 0x7d, 0x24, 0xfe, 0xe9,
-            0x12, 0x19, 0xf, 0xc2, 0x73, 0x1c, 0xe3, 0x76, 0x91, 0xec, 0x57, 0x6c, 0xcd, 0x7b,
-            0xab, 0x32, 0xfd, 0x6d, 0x6e, 0x92, 0x7d, 0x37, 0x60, 0x1, 0xdb, 0x13, 0x92, 0x3b,
-            0x77, 0xf7, 0x12, 0x97, 0x1d, 0x5e, 0xe3, 0xb9, 0x15, 0x83, 0xaf, 0x89, 0xa3, 0x81,
-            0x9c, 0x30, 0x81, 0x99, 0x30, 0xc, 0x6, 0x3, 0x55, 0x1d, 0x13, 0x1, 0x1, 0xff, 0x4,
-            0x2, 0x30, 0x0, 0x30, 0xb, 0x6, 0x3, 0x55, 0x1d, 0xf, 0x4, 0x4, 0x3, 0x2, 0x5, 0xe0,
-            0x30, 0x1d, 0x6, 0x3, 0x55, 0x1d, 0xe, 0x4, 0x16, 0x4, 0x14, 0x48, 0x1f, 0x5d, 0x95,
-            0xce, 0x89, 0xd4, 0x7d, 0xa4, 0x4c, 0x21, 0x8f, 0x5b, 0xd5, 0x50, 0x96, 0xff, 0xba,
-            0xe2, 0xee, 0x30, 0x31, 0x6, 0x3, 0x55, 0x1d, 0x11, 0x4, 0x2a, 0x30, 0x28, 0xa0, 0x26,
-            0x6, 0xa, 0x2b, 0x6, 0x1, 0x4, 0x1, 0x83, 0x1c, 0x82, 0x12, 0x1, 0xa0, 0x18, 0xc, 0x16,
-            0x41, 0x43, 0x4d, 0x45, 0x3a, 0x57, 0x49, 0x44, 0x47, 0x45, 0x54, 0x3a, 0x31, 0x32,
-            0x33, 0x34, 0x35, 0x36, 0x37, 0x38, 0x39, 0x30, 0x30, 0x2a, 0x6, 0x3, 0x55, 0x1d, 0x25,
-            0x1, 0x1, 0xff, 0x4, 0x20, 0x30, 0x1e, 0x6, 0x8, 0x2b, 0x6, 0x1, 0x5, 0x5, 0x7, 0x3,
-            0x1, 0x6, 0x8, 0x2b, 0x6, 0x1, 0x5, 0x5, 0x7, 0x3, 0x2, 0x6, 0x8, 0x2b, 0x6, 0x1, 0x5,
-            0x5, 0x7, 0x3, 0x9, 0x30, 0xa, 0x6, 0x8, 0x2a, 0x86, 0x48, 0xce, 0x3d, 0x4, 0x3, 0x3,
-            0x3, 0x68, 0x0, 0x30, 0x65, 0x2, 0x30, 0x8, 0xe6, 0x1f, 0xd, 0xdf, 0x18, 0xd3, 0x2f,
-            0x50, 0x49, 0x99, 0xb0, 0xe2, 0x64, 0x95, 0x30, 0xa9, 0x5a, 0xbf, 0x83, 0x76, 0xae,
-            0x4a, 0x39, 0xd8, 0xe2, 0x51, 0x12, 0x84, 0x9c, 0xbe, 0x11, 0x1d, 0x3b, 0x77, 0x20,
-            0x6f, 0x5, 0x6c, 0xc7, 0x98, 0xb2, 0xba, 0xb8, 0x96, 0x75, 0x25, 0xcf, 0x2, 0x31, 0x0,
-            0x93, 0x12, 0x5b, 0x66, 0x93, 0xc0, 0xe7, 0x56, 0x1b, 0x68, 0x28, 0x27, 0xd8, 0x8e,
-            0x69, 0xaa, 0x30, 0x76, 0x5, 0x6f, 0x4b, 0xd0, 0xce, 0x10, 0xf, 0xf8, 0xdf, 0x4a, 0xab,
-            0x9b, 0x4d, 0xb1, 0x47, 0xe4, 0xcd, 0xce, 0xce, 0x48, 0xd, 0xf8,
-        ],
-        [
-            0x1, 0x0, 0x1, 0x0, 0x6, 0x0, 0x0, 0x0, 0x11, 0x2, 0x0, 0x0, 0x8, 0x0, 0x0, 0x0, 0x35,
-            0x3d, 0xbc, 0x25, 0xce, 0xec, 0xb9, 0xca, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-            0x0, 0x0,
-        ],
-    ]
-}
-
-pub fn get_rsp_cert_chain_buff() -> SpdmCertChainBuffer {
-    use spdmlib::crypto;
-    let hash_algo = SpdmBaseHashAlgo::TPM_ALG_SHA_384;
-    let cert_chain = include_bytes!("../../../test_key/ecp384/bundle_responder.certchain.der");
-
-    let (root_cert_begin, root_cert_end) =
-        crypto::cert_operation::get_cert_from_cert_chain(cert_chain, 0)
-            .expect("Get provisioned root cert failed");
-
-    let root_cert_hash =
-        crypto::hash::hash_all(hash_algo, &cert_chain[root_cert_begin..root_cert_end])
-            .expect("Must provide hash algo");
-    SpdmCertChainBuffer::new(cert_chain, root_cert_hash.as_ref())
-        .expect("Create format certificate chain failed.")
-}
