--- fuzz-target/fuzzlib/src/shared_buffer.rs
@@ -1,40 +1,0 @@
-// Copyright (c) 2021 Intel Corporation
-//
-// SPDX-License-Identifier: BSD-2-Clause-Patent
-
-extern crate alloc;
-use alloc::collections::VecDeque;
-use core::cell::RefCell;
-
-pub struct SharedBuffer {
-    queue: RefCell<VecDeque<u8>>,
-}
-
-impl SharedBuffer {
-    pub fn new() -> Self {
-        SharedBuffer {
-            queue: RefCell::new(VecDeque::<u8>::new()),
-        }
-    }
-    pub fn set_buffer(&self, b: &[u8]) {
-        log::info!("send    {:02x?}\n", b);
-        let mut queue = self.queue.borrow_mut();
-        for i in b {
-            queue.push_back(*i);
-        }
-    }
-
-    pub fn get_buffer(&self, b: &mut [u8]) -> usize {
-        let mut queue = self.queue.borrow_mut();
-        let mut len = 0usize;
-        for i in b.iter_mut() {
-            if queue.is_empty() {
-                break;
-            }
-            *i = queue.pop_front().unwrap();
-            len += 1;
-        }
-        log::info!("recieve {:02x?}\n", &b[..len]);
-        len
-    }
-}
