+++ fuzz-target/pass_context/src/pass_responder.rs
@@ -8,7 +8,6 @@ pub fn pass_rsp_handle_spdm_version() {
     let (config_info, provision_info) = rsp_create_info();
 
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -17,11 +16,7 @@ pub fn pass_rsp_handle_spdm_version() {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
@@ -35,7 +30,6 @@ pub fn pass_rsp_handle_spdm_version() {
 pub fn pass_rsp_handle_spdm_capability() {
     let (config_info, provision_info) = rsp_create_info();
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -44,11 +38,7 @@ pub fn pass_rsp_handle_spdm_capability() {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
@@ -63,7 +53,6 @@ pub fn pass_rsp_handle_spdm_capability() {
 pub fn pass_rsp_handle_spdm_algorithm() {
     let (config_info, provision_info) = rsp_create_info();
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -72,11 +61,7 @@ pub fn pass_rsp_handle_spdm_algorithm() {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
@@ -93,7 +78,6 @@ pub fn pass_rsp_handle_spdm_algorithm() {
 pub fn pass_rsp_handle_spdm_digest() {
     let (config_info, provision_info) = rsp_create_info();
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -102,11 +86,7 @@ pub fn pass_rsp_handle_spdm_digest() {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
@@ -125,7 +105,6 @@ pub fn pass_rsp_handle_spdm_digest() {
 pub fn pass_rsp_handle_spdm_certificate() {
     let (config_info, provision_info) = rsp_create_info();
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -134,11 +113,7 @@ pub fn pass_rsp_handle_spdm_certificate() {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
@@ -156,7 +131,6 @@ pub fn pass_rsp_handle_spdm_certificate() {
 pub fn pass_rsp_handle_spdm_challenge() {
     let (config_info, provision_info) = rsp_create_info();
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -165,11 +139,7 @@ pub fn pass_rsp_handle_spdm_challenge() {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
@@ -191,7 +161,6 @@ pub fn pass_rsp_handle_spdm_challenge() {
 pub fn pass_rsp_handle_spdm_measurement() {
     let (config_info, provision_info) = rsp_create_info();
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -200,11 +169,7 @@ pub fn pass_rsp_handle_spdm_measurement() {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
@@ -227,7 +192,6 @@ pub fn pass_rsp_handle_spdm_measurement() {
 pub fn pass_rsp_handle_spdm_key_exchange() {
     let (config_info, provision_info) = rsp_create_info();
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -236,11 +200,7 @@ pub fn pass_rsp_handle_spdm_key_exchange() {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
@@ -274,7 +234,6 @@ pub fn pass_rsp_handle_spdm_key_exchange() {
 pub fn pass_rsp_handle_spdm_psk_exchange() {
     let (config_info, provision_info) = rsp_create_info();
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -283,11 +242,7 @@ pub fn pass_rsp_handle_spdm_psk_exchange() {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
