+++ fuzz-target/requester/algorithm_req/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2018"
 [dependencies]
 fuzzlib = { path = "../../fuzzlib", default-features = false }
 afl = { version = "=0.12.12", optional = true }
+spdmlib-test = { path = "../../../test/spdmlib-test" }
 
 [features]
 fuzzlogfile = []
