+++ fuzz-target/requester/psk_exchange_req/src/main.rs
@@ -8,10 +8,8 @@ use spdmlib::protocol::*;
 
 fn fuzz_send_receive_spdm_psk_exchange(fuzzdata: &[u8]) {
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
-    spdmlib::secret::measurement::register(
-        fuzzlib::secret::SECRET_MEASUREMENT_IMPL_INSTANCE.clone(),
-    );
-    spdmlib::secret::psk::register(fuzzlib::secret::SECRET_PSK_IMPL_INSTANCE.clone());
+    spdmlib::secret::measurement::register(SECRET_MEASUREMENT_IMPL_INSTANCE.clone());
+    spdmlib::secret::psk::register(SECRET_PSK_IMPL_INSTANCE.clone());
     spdmlib::crypto::hkdf::register(FAKE_HKDF.clone());
 
     // TCD:
