+++ fuzz-target/requester/psk_finish_req/src/main.rs
@@ -10,7 +10,7 @@ use spdmlib::protocol::*;
 
 fn fuzz_send_receive_spdm_psk_finish(fuzzdata: &[u8]) {
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
-    spdmlib::secret::psk::register(fuzzlib::secret::SECRET_PSK_IMPL_INSTANCE.clone());
+    spdmlib::secret::psk::register(SECRET_PSK_IMPL_INSTANCE.clone());
     spdmlib::crypto::aead::register(FAKE_AEAD.clone());
 
     let (req_config_info, req_provision_info) = req_create_info();
