+++ fuzz-target/responder/algorithm_rsp/src/main.rs
@@ -8,7 +8,6 @@ use spdmlib::common::SpdmConnectionState;
 fn fuzz_handle_spdm_algorithm(data: &[u8]) {
     let (config_info, provision_info) = rsp_create_info();
     let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
-    let mctp_transport_encap = &mut MctpTransportEncap {};
 
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
 
@@ -17,11 +16,7 @@ fn fuzz_handle_spdm_algorithm(data: &[u8]) {
 
     let mut context = responder::ResponderContext::new(
         &mut socket_io_transport,
-        if USE_PCIDOE {
-            pcidoe_transport_encap
-        } else {
-            mctp_transport_encap
-        },
+        pcidoe_transport_encap,
         config_info,
         provision_info,
     );
