+++ fuzz-target/responder/keyexchange_rsp/src/main.rs
@@ -8,9 +8,7 @@ use spdmlib::protocol::*;
 
 fn fuzz_handle_spdm_key_exchange(data: &[u8]) {
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
-    spdmlib::secret::measurement::register(
-        fuzzlib::secret::SECRET_MEASUREMENT_IMPL_INSTANCE.clone(),
-    );
+    spdmlib::secret::measurement::register(SECRET_MEASUREMENT_IMPL_INSTANCE.clone());
 
     let (config_info, provision_info) = rsp_create_info();
     let shared_buffer = SharedBuffer::new();
