+++ fuzz-target/responder/psk_finish_rsp/src/main.rs
@@ -10,7 +10,7 @@ use spdmlib::protocol::*;
 
 fn fuzz_handle_spdm_psk_finish(data: &[u8]) {
     spdmlib::secret::asym_sign::register(SECRET_ASYM_IMPL_INSTANCE.clone());
-    spdmlib::secret::psk::register(fuzzlib::secret::SECRET_PSK_IMPL_INSTANCE.clone());
+    spdmlib::secret::psk::register(SECRET_PSK_IMPL_INSTANCE.clone());
     spdmlib::crypto::hmac::register(FAKE_HMAC.clone());
     spdmlib::crypto::hkdf::register(FAKE_HKDF.clone());
